Rename clientinfoutils to clientinfo
Also switch to using trace.Log(ctx) in clientinfo
